refactor: accept remoteOptions as a param to all remote APIs, prevent from passing lot of the similar params to the similar APIs constructors.
